requirements.txt
Django

# Schema migrations
south

# For the admin documentation
docutils

# For the Imagefields
Pillow

# Async jobs
django-celery>=3.0.11

# Timezonefoo
pytz>dev

# Gravatar
django-gravatar2

# HTML minifying
django-htmlmin

# AJAX & API
djangorestframework

# Authentication
#python-openid
#django-allauth

# Caching
python-memcached

# Database
psycopg2

# Twitter cards
git+git://github.com/leveille/django-twittercard.git#egg=twittercard

# GnuPG for outgoing post mails
python-gnupg
django-email-extras

# API authentication
django-oauth2-provider

# Language detection
guess-language
babel

# Stripe payments
stripe

